Refactor: bounded 부분 modifier 확장함수 분리
#PORRING-100
