+++ feature/image-edit/src/main/java/com/kolown/porring/feature/imageedit/ImageEditScreen.kt
@@ -57,6 +57,7 @@ import com.kolown.porring.core.designsystem.component.PorringIconButton
 import com.kolown.porring.core.designsystem.component.PorringTopAppBar
 import com.kolown.porring.core.designsystem.ui.theme.Primary
 import com.kolown.porring.core.ui.R.drawable
+import com.kolown.porring.feature.imageedit.component.boundedTransformGestures
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
@@ -176,73 +177,15 @@ private fun ImageEditScreen(
                         }
                     }
                 }
-                .pointerInput(boxSize, imageSize, minScale) {
-                    awaitPointerEventScope {
-                        while (true) {
-                            val event = awaitPointerEvent()
-
-                            if (event.type == PointerEventType.Release) {
-                                coroutineScope.launch {
-                                    if (scale.value < minScale) {
-                                        val scaleJob = async {
-                                            scale.animateTo(
-                                                minScale,
-                                                animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                            )
-                                        }
-                                        val offsetXJob = async {
-                                            offsetX.animateTo(
-                                                0f,
-                                                animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                            )
-                                        }
-                                        val offsetYJob = async {
-                                            offsetY.animateTo(
-                                                0f,
-                                                animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                            )
-                                        }
-
-                                        awaitAll(scaleJob, offsetXJob, offsetYJob)
-                                    }
-
-                                    if (scale.value > 5f) {
-                                        scale.animateTo(
-                                            5f,
-                                            animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                        )
-                                    }
-
-                                    val boundedOffset = getBoundedOffset(
-                                        scale.value,
-                                        boxSize,
-                                        imageSize,
-                                        Offset(offsetX.value, offsetY.value)
-                                    )
-
-                                    val offsetXJob = async {
-                                        offsetX.animateTo(
-                                            boundedOffset.x,
-                                            animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                        )
-                                    }
-
-                                    val offsetYJob = async {
-                                        offsetY.animateTo(
-                                            boundedOffset.y,
-                                            animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy)
-                                        )
-                                    }
-
-                                    awaitAll(
-                                        offsetXJob,
-                                        offsetYJob
-                                    )
-                                }
-                            }
-                        }
-                    }
-                },
+                .boundedTransformGestures(
+                    coroutineScope = coroutineScope,
+                    scale = scale,
+                    offsetX = offsetX,
+                    offsetY = offsetY,
+                    minScale = minScale,
+                    boxSize = boxSize,
+                    imageSize = imageSize
+                ),
             contentAlignment = Alignment.Center
         ) {
             AsyncImage(
@@ -359,24 +302,6 @@ private fun calcMinScale(boxSize: Size, imageSize: Size): Float {
     val (boxW, boxH) = boxSize
     val (imgW, imgH) = imageSize
     return if ((imgW / imgH) > (boxW / boxH)) boxH / imgH else boxW / imgW
-}
-
-private fun getBoundedOffset(
-    scale: Float,
-    boxSize: Size,
-    imageSize: Size,
-    currentOffset: Offset
-): Offset {
-    val scaledWidth = imageSize.width * scale
-    val scaledHeight = imageSize.height * scale
-
-    val maxOffsetX = ((scaledWidth - boxSize.width) / 2f).coerceAtLeast(0f)
-    val maxOffsetY = ((scaledHeight - boxSize.height) / 2f).coerceAtLeast(0f)
-
-    return Offset(
-        currentOffset.x.coerceIn(-maxOffsetX, maxOffsetX),
-        currentOffset.y.coerceIn(-maxOffsetY, maxOffsetY)
-    )
 }
 
 enum class CropRatio(val ratio: Float) { PORTRAIT(4f / 5f), LANDSCAPE(5f / 4f) }
